requirements.txt
speedtest-cli
pandas
pymongo
